# Improve keccak cuda performance 35%