--- a/risc0/circuit/keccak/src/prove/tests.rs
+++ b/risc0/circuit/keccak/src/prove/tests.rs
@@ -27,7 +27,10 @@ use test_log::test;
 use super::{testutil::test_inputs, zkr::get_keccak_zkr};
 use crate::{
     get_control_id,
-    prove::{keccak_prover, CircuitWitnessGenerator as _, MetaBuffer, PreflightTrace, StepMode},
+    prove::{
+        keccak_prover, preflight::ForwardPreflightOrder, CircuitWitnessGenerator as _, MetaBuffer,
+        PreflightTrace, StepMode,
+    },
     zirgen::circuit::*,
     KECCAK_PO2_RANGE,
 };
@@ -61,7 +64,7 @@ fn fwd_rev_ab() {
     let po2 = 8;
     let inputs = test_inputs(po2);
     let cycles: usize = 1 << po2;
-    let preflight = PreflightTrace::new(&inputs, cycles);
+    let preflight = PreflightTrace::<ForwardPreflightOrder>::new(&inputs, cycles);
 
     let fwd_data = {
         let global = MetaBuffer::new("global", hal.as_ref(), 1, REGCOUNT_GLOBAL, true);