* Reorder how the witness is generated to cut down on warp divergence

Before:

```
10 runs of PO2=16 completed in 6.543s, avg=0.654s, 100157.742 cycles/sec, 499.749 keccak/sec
10 runs of PO2=17 completed in 11.796s, avg=1.180s, 111115.459 cycles/sec, 555.272 keccak/sec
```

After:

```
10 runs of PO2=16 completed in 4.247s, avg=0.425s, 154322.577 cycles/sec, 770.012 keccak/sec
10 runs of PO2=17 completed in 7.548s, avg=0.755s, 173643.059 cycles/sec, 867.738 keccak/sec
```